#5: added unit tests
