+++ app/src/main/java/com/example/concordia_campus_guide/InfoCardFragment/InfoCardFragmentViewModel.java
@@ -6,6 +6,7 @@ import android.widget.TextView;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.ViewModel;
 
 import com.example.concordia_campus_guide.Models.Buildings;
 import com.example.concordia_campus_guide.R;
@@ -15,18 +16,11 @@ import java.io.IOException;
 import java.io.InputStream;
 
 // This class is for the business logic
-public class InfoCardFragmentViewModel extends AndroidViewModel {
+public class InfoCardFragmentViewModel extends ViewModel {
 
     private Buildings buildings;
     private TextView infoCardTitle;
     private TextView buildingAddress;
-
-    /**
-     * Constructor
-     */
-    public InfoCardFragmentViewModel(@NonNull Application application) {
-        super(application);
-    }
 
     /**
      * Reads json file and outputs Buildings object, containing many Building objects
@@ -39,7 +33,7 @@ public class InfoCardFragmentViewModel extends AndroidViewModel {
         Buildings buildings = new Buildings();
 
         try{
-            InputStream is = getApplication().getResources().openRawResource(R.raw.buildings_info);
+            InputStream is = context.getResources().openRawResource(R.raw.buildings_info);
             int size = is.available();
             byte[] buffer = new byte[size];
             is.read(buffer);
